services/scheduler/pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = "Resource Scheduler for 5G IoT Edge Gateway — decides edge vs cloud placement for ML inference tasks"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "grpcio>=1.60.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scheduler*"]

[project.scripts]
scheduler = "scheduler.main:run"

# Install the shared library as a path dependency:
#   pip install -e ../../shared
# Or in pip-tools / Docker:
#   pip install ../../shared
[tool.scheduler.dependencies-path]
shared = {path = "../../shared"}
